major code refactoring, install using pip, --area to find servers in a city,state,area. full country names supported.
