+++ .gitignore
@@ -1,4 +1,8 @@
-pass*
+__pycache__
+openpyn.egg*
+build
+dist
 files/*
+openpyn/files
 zip
 creds
