+++ openpyn/__init__.py
@@ -1,0 +1,2 @@
+__version__ = "1.4.1"
+__license__ = "GNU General Public License v3 or later (GPLv3+)"
