+++ setup.cfg
@@ -1,0 +1,5 @@
+[bdist_wheel]
+
+[metadata]
+description-file = README.rst
+license_file = LICENSE.md
